lala

+++ s0vz.c
@@ -265,7 +265,7 @@ void update_average_values(struct valuePack *vP) {
 	{
 		int time = (int)ts-vP->lastTs;
 		double wattProImpuls = 1000.0 / (double)vP->impulsConst;
-        double tmp_value = wattProImpuls * (3.6 / time) * 1000000.0; // Zeit in MS
+        double tmp_value = wattProImpuls * (3.6 / (double)time) * 1000000.0; // Zeit in MS
 	    vP->valuesAsSumm += tmp_value / 1000.0;
 	    vP->numberOfValues++;
 	    printf("Summe: %.3f Anzahl %d TMPValue: %.3f Watt/imp: %f \n", vP->valuesAsSumm, vP->numberOfValues, tmp_value, wattProImpuls );
